moved shuffle to utils/helpers
