+++ src/App.js
@@ -1,6 +1,7 @@
 import './App.css';
-import React from "react";
-import Card from "../src/components/cards/Card.js"
+import React from 'react';
+import Card from './components/cards/Card.js';
+import shuffle from './utils/helpers/shuffle';
 
 function App() {
   const cardContent = {
@@ -8,17 +9,6 @@ function App() {
     sol: "Não poderia ser diferente a mensagem quando o assunto é o O Sol. Sua luz é sinal de positividade e está completamente associado aos objetivos e sucesso.",
     mundo: "É aqui que aparece as conquistas feitas, os aprendizados adquiridos, as barreiras rompidas e toda a batalha travada.",
     estrela: "Não é de se admirar que a estrela represente em nossas vidas sinal de luz e renovação. Sem essa âncora é possível agora caminhar em direção a uma nova vida."
-  }
-
-  function shuffle(array) {
-    const copy = [...array];
-    const result = [];
-
-    while (copy.length > 0) {
-      result.push(copy.splice(Math.floor(Math.random() * copy.length), 1));
-    }
-
-    return result;
   }
 
   const shuffledCards = shuffle(Object.entries(cardContent))
